# Fix reported time for multithreaded benchmarks