In report real and manual per iteration is divided by thread count
because total time is adjusted to per-thread while iterations counter is not.
See #769 

Process CPU time is special - each thread reports time spent by all threads
thus it actually must be divided by thread count once to get total cpu time and
second time to get per-thread CPU time.