@LebedevRI 2020-03-15
Thank you for looking into this. With multi-threaded benchmarks, It is hard to tell whether
the currently-reported numbers are correct or not, as noted in mentioned #769.