@koct9i 2020-03-15
Bug in current reports is pretty obvious for trivial benchmark which spins or sleep fixed amount of time.
It seems present test/internal_threading_test.cc could be used for that. But it needs validation.
But writing checks for it using "output_test" is a pain =) I'll play with custom reported, this should easier.