@koct9i 2020-03-31
Oh, I get it - spin-wait in test watching into real time. :man_facepalming: