@dominichamon 2020-06-17
the code seems reasonable. The threads division has changed back and forth a few times (as i mentioned in #769).

i'm just running a diff between the head version and this change to see what the output looks like for tests.