@dominichamon 2020-06-17
This is the before

```
----------------------------------------------------------------------------------------
Benchmark                                              Time             CPU   Iterations
----------------------------------------------------------------------------------------
BM_spin_empty/8/threads:4                           23.0 ns         90.2 ns      7967324
BM_spin_empty/512/threads:4                         1367 ns         4908 ns       158756
BM_spin_empty/8192/threads:4                       22378 ns        79529 ns         9364
```

This is the after
```
----------------------------------------------------------------------------------------
Benchmark                                              Time             CPU   Iterations
----------------------------------------------------------------------------------------
BM_spin_empty/8/threads:4                            168 ns         95.0 ns     33131036
BM_spin_empty/512/threads:4                         9492 ns         4960 ns       512840
BM_spin_empty/8192/threads:4                      155608 ns        80357 ns        31952
```

Before this change, it's clear that the time is the end-to-end time given the parallelisation (23 ~= 90/4, 1300 ~= 4900/4, etc). In the latter case i find it hard to understand why we're ~doubling the reported time.

another example:

before
```
BM_pause_during                                      804 ns          796 ns       798884
BM_pause_during/threads:4                            355 ns         1221 ns       568668
BM_pause_during/real_time                            972 ns          970 ns       848250
BM_pause_during/real_time/threads:4                  361 ns         1250 ns      1890216
```

after
```
BM_pause_during                                     1248 ns         1188 ns       594011
BM_pause_during/threads:4                           2873 ns         1297 ns      2099924
BM_pause_during/real_time                           1326 ns         1196 ns       552044
BM_pause_during/real_time/threads:4                 2357 ns         1300 ns       975240
```

again i find it hard to understand why the threading appears to cause the measured time to double. the former is what i'd expect: ~4x improvement by using 4 threads.