@koct9i 2020-06-18
My patch normalize all times to 1 iteration in 1 thread. So, if threads runs all time then "real" time should be equal to 'CPU" time or be slightly more. For perfect parallelism times should stay the same regardless of count of threads. 

In your case 'real' time much bigger then 'cpu', this means threads were preempted. I guess you are using osx? Probably these artifacts were caused by it's scheduler.

For Linux:

Before
```
BM_spin_empty/8                                     43.2 ns         43.2 ns     16180998
BM_spin_empty/512                                   2448 ns         2448 ns       285791
BM_spin_empty/8192                                 39142 ns        39138 ns        17843
BM_spin_empty/8/threads:8                           11.8 ns         94.0 ns      7782976
BM_spin_empty/512/threads:8                          708 ns         5638 ns       125112
BM_spin_empty/8192/threads:8                       11666 ns        92589 ns         7576
```

After
```
BM_spin_empty/8                                     46.8 ns         46.8 ns     14851377
BM_spin_empty/512                                   2638 ns         2638 ns       263970
BM_spin_empty/8192                                 42235 ns        42222 ns        16626
BM_spin_empty/8/threads:8                            109 ns          109 ns     54453880
BM_spin_empty/512/threads:8                         6212 ns         6201 ns       911128
BM_spin_empty/8192/threads:8                      101850 ns       101688 ns        55528
```

As you can see in 'Before' 'real' time was erroneously  divided by 8.

Performance for per-cpu threads is around twice lower because this is 4 core cpu with hyper-threading and turbo-boost cannot reach top frequency when all cores are active.