@dominichamon 2020-06-18
I think this change is correct. The wall-time (Time) and CPU time should match in terms of per-thread vs global. The benefit of threading should be visible when the real and CPU times don't change but the amount of processing (items per second, etc) increases. As noted, if threading _does_ lead to a change in timing it indicates where the parallelism breaks down.