@krzikalla 2020-08-14
Since it is related, I want to add another problem to this issue: the real time currently measures the average wall-clock time  of every single thread. It should rather measures the wall-clock time of the thread pack. That is, the actual real time is the time, when the last thread enters FinishKeepRunning. Only by that means you can spot balancing issues.
Example: if one thread runs 1 sec and 7 threads run 1 ms each, then currently 0,125 ms real time is reported. This is not the wall-clock time! The wall-clock time is 1 sec.
Solution: one could move the barrier in FinishKeepRunning before PauseTiming. But this would tamper the timing. Instead don't sum up the real times, but use the maximum one at https://github.com/google/benchmark/blob/1302d2ce094a9753b0f81a81ea74c0fa71fae582/src/benchmark_runner.cc#L127. Of course you must not divide real time by num_threads then.