@koct9i 2020-08-18
@krzikalla 
> Example: if one thread runs 1 sec and 7 threads run 1 ms each, then currently 0,125 sec real time is reported. This is not the wall-clock time! The wall-clock time is 1 sec.

This seems odd. If each thread did same single iteration then this means execution speed is unpredictable.
Instead of forging "sane" metrics benchmark should measure and report deviation among iterations.

Present design of benchmark loop could easily estimate deviation among averages for each thread.
To catch variations within thread it have to measure time periodically or compare progress.