@krzikalla 2020-08-24
> would it be ok to defer the wall-clock time change to a separate PR or do you want to solve that as part of this change?

It can be a separate PR. I just wanted to raise awareness for this issue and thought, it would be better to add it to this discussion here instead of creating a new one.

> This seems odd. If each thread did same single iteration then this means execution speed is unpredictable.

Not necessarily. Other possible scenarios are poor balancing or a hit by Amdahl's law.

> Instead of forging "sane" metrics benchmark should measure and report deviation among iterations.
> 
> Present design of benchmark loop could easily estimate deviation among averages for each thread.
> To catch variations within thread it have to measure time periodically or compare progress.

Normally, you don't report deviations. You report wall clock time, since that affect the usability of the code.